Big refactoring, address comments in PR
